requirements.txt
Django
Pillow
djangorestframework
django-dirtyfields
psycopg2
django-reversion
kombu
celery
python-dateutil
lxml
six
requests
-e git://github.com/alexsdutton/idm-brand.git#egg=idm-brand
-e git://github.com/alexsdutton/idm-broker.git#egg=idm-notification
-e git://github.com/alexsdutton/django-oidc-auth.git#egg=django-oidc-auth
pyjwkest>=1.3,<1.4
